# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PNAS 项目 API 文档",
        "title": "PNAS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "description": "用户登录获取JWT Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "$ref": "#/definitions/v1.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户名或密码错误",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户登出（使Token失效）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登出",
                "responses": {
                    "200": {
                        "description": "登出成功",
                        "schema": {
                            "$ref": "#/definitions/v1.LogoutResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "description": "使用现有Token获取新的Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "刷新Token",
                "parameters": [
                    {
                        "description": "刷新Token请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.RefreshTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token刷新成功",
                        "schema": {
                            "$ref": "#/definitions/v1.RefreshTokenResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "无效的Token",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.UserStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-comments": {
                "UserStatusActive": "正常",
                "UserStatusDisabled": "禁用",
                "UserStatusLocked": "锁定"
            },
            "x-enum-descriptions": [
                "禁用",
                "正常",
                "锁定"
            ],
            "x-enum-varnames": [
                "UserStatusDisabled",
                "UserStatusActive",
                "UserStatusLocked"
            ]
        },
        "models.UserType": {
            "type": "integer",
            "enum": [
                1,
                2,
                3,
                4
            ],
            "x-enum-comments": {
                "UserTypeAudit": "审计管理员",
                "UserTypeNormal": "普通用户",
                "UserTypeSecurity": "安全管理员",
                "UserTypeSystem": "系统管理员"
            },
            "x-enum-descriptions": [
                "系统管理员",
                "安全管理员",
                "审计管理员",
                "普通用户"
            ],
            "x-enum-varnames": [
                "UserTypeSystem",
                "UserTypeSecurity",
                "UserTypeAudit",
                "UserTypeNormal"
            ]
        },
...
        "v1.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "validation failed"
                },
                "message": {
                    "type": "string",
                    "example": "请求参数错误"
                }
            }
        },
        "v1.LoginData": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string",
                    "example": "2025-08-08T23:59:59Z"
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user": {
                    "$ref": "#/definitions/v1.UserInfo"
                }
            }
        },
        "v1.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "password123"
                },
                "username": {
                    "type": "string",
                    "example": "sysadmin"
                }
            }
        },
        "v1.LoginResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/v1.LoginData"
                },
                "message": {
                    "type": "string",
                    "example": "登录成功"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "v1.LogoutResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "登出成功"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "v1.RefreshTokenData": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string",
                    "example": "2025-08-08T23:59:59Z"
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
        "v1.RefreshTokenRequest": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
        "v1.RefreshTokenResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/v1.RefreshTokenData"
                },
                "message": {
                    "type": "string",
                    "example": "Token刷新成功"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "v1.UserGroupInfo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "负责系统配置、用户管理、系统维护"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "系统管理员组"
                }
            }
        },
...
        "v1.UserInfo": {
            "type": "object",
            "properties": {
                "group": {
                    "$ref": "#/definitions/v1.UserGroupInfo"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.UserStatus"
                        }
                    ],
                    "example": 1
                },
                "user_type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.UserType"
                        }
                    ],
                    "example": 1
                },
                "username": {
                    "type": "string",
                    "example": "sysadmin"
                }
            }
        },
...
    }
}